+++ api/src/Geography/Entity/Address.php
@@ -50,7 +50,7 @@ use Doctrine\Common\Collections\ArrayCollection;
  * @ApiResource(
  *      attributes={
  *          "force_eager"=false,
- *          "normalization_context"={"groups"={"read","pt","mass","search"}, "enable_max_depth"="true"},
+ *          "normalization_context"={"groups"={"read","pt","mass","search","readRelayPoint"}, "enable_max_depth"="true"},
  *          "denormalization_context"={"groups"={"write"}}
  *      },
  *      collectionOperations={
@@ -102,7 +102,10 @@ use Doctrine\Common\Collections\ArrayCollection;
  *              "security"="is_granted('import_create',object)"
  *          },
  *      },
- *      itemOperations={"get","put"}
+ *      itemOperations={
+ *          "get",
+ *          "put"
+ *      }
  * )
  * @ApiFilter(OrderFilter::class, properties={"id", "streetAddress", "postalCode", "addressLocality", "addressCountry"}, arguments={"orderParameterName"="order"})
  */
@@ -118,7 +121,7 @@ class Address implements \JsonSerializable
      * @ORM\Id
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
-     * @Groups({"read", "readUser", "readEvent"})
+     * @Groups({"read", "readUser", "readEvent", "readRelayPoint"})
      * @ApiProperty(identifier=true)
      */
     private $id;
@@ -127,7 +130,7 @@ class Address implements \JsonSerializable
      * @var string The house number.
      *
      * @ORM\Column(type="string", length=45, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $houseNumber;
 
@@ -135,7 +138,7 @@ class Address implements \JsonSerializable
      * @var string The street.
      *
      * @ORM\Column(type="string", length=255, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","readRelayPoint"})
      * @Assert\NotBlank(groups={"mass","massCompute","threads","thread"})
      */
     private $street;
@@ -144,7 +147,7 @@ class Address implements \JsonSerializable
      * @var string The full street address.
      *
      * @ORM\Column(type="string", length=255, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","externalJourney"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","externalJourney","readRelayPoint"})
      */
     private $streetAddress;
 
@@ -152,7 +155,7 @@ class Address implements \JsonSerializable
      * @var string|null The postal code of the address.
      *
      * @ORM\Column(type="string", length=15, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","externalJourney"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","externalJourney","readRelayPoint"})
      * @Assert\NotBlank(groups={"mass","massCompute","threads","thread"})
      */
     private $postalCode;
@@ -161,7 +164,7 @@ class Address implements \JsonSerializable
      * @var string|null The sublocality of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $subLocality;
 
@@ -169,7 +172,7 @@ class Address implements \JsonSerializable
      * @var string|null The locality of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","readEvent","results","write","pt","mass","massCompute","threads","thread","externalJourney","readCommunity"})
+     * @Groups({"read","readUser","readEvent","results","write","pt","mass","massCompute","threads","thread","externalJourney","readCommunity","readRelayPoint"})
      * @Assert\NotBlank(groups={"mass","massCompute","threads","thread"})
      */
     private $addressLocality;
@@ -178,7 +181,7 @@ class Address implements \JsonSerializable
      * @var string|null The locality admin of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $localAdmin;
 
@@ -186,7 +189,7 @@ class Address implements \JsonSerializable
      * @var string|null The county of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $county;
 
@@ -194,7 +197,7 @@ class Address implements \JsonSerializable
      * @var string|null The macro county of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $macroCounty;
 
@@ -202,7 +205,7 @@ class Address implements \JsonSerializable
      * @var string|null The region of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $region;
 
@@ -210,7 +213,7 @@ class Address implements \JsonSerializable
      * @var string|null The macro region of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $macroRegion;
 
@@ -218,7 +221,7 @@ class Address implements \JsonSerializable
      * @var string|null The country of the address.
      *
      * @ORM\Column(type="string", length=100, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","externalJourney"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","externalJourney","readRelayPoint"})
      */
     private $addressCountry;
 
@@ -226,7 +229,7 @@ class Address implements \JsonSerializable
      * @var string|null The country code of the address.
      *
      * @ORM\Column(type="string", length=10, nullable=true)
-     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread"})
+     * @Groups({"read","readUser","results","write","pt","mass","massCompute","threads","thread","readRelayPoint"})
      */
     private $countryCode;
 
@@ -234,7 +237,7 @@ class Address implements \JsonSerializable
      * @var float|null The latitude of the address.
      *
      * @ORM\Column(type="decimal", precision=10, scale=6, nullable=true)
-     * @Groups({"read","readUser","readCommunity","readEvent","results","write","pt","mass","massCompute","threads","thread","externalJourney"})
+     * @Groups({"read","readUser","readCommunity","readEvent","results","write","pt","mass","massCompute","threads","thread","externalJourney","readRelayPoint"})
      */
     private $latitude;
 
@@ -242,7 +245,7 @@ class Address implements \JsonSerializable
      * @var float|null The longitude of the address.
      *
      * @ORM\Column(type="decimal", precision=10, scale=6, nullable=true)
-     * @Groups({"read","readUser","readCommunity","readEvent","results","write","pt","mass","massCompute","threads","thread","externalJourney"})
+     * @Groups({"read","readUser","readCommunity","readEvent","results","write","pt","mass","massCompute","threads","thread","externalJourney","readRelayPoint"})
      */
     private $longitude;
 
@@ -250,7 +253,7 @@ class Address implements \JsonSerializable
      * @var int|null The elevation of the address in metres.
      *
      * @ORM\Column(type="integer", nullable=true)
-     * @Groups({"read","results","write","pt","mass","massCompute"})
+     * @Groups({"read","results","write","pt","mass","massCompute","readRelayPoint"})
      */
     private $elevation;
 
@@ -273,7 +276,7 @@ class Address implements \JsonSerializable
      * @var string|null The venue name of this address.
      *
      * @ORM\Column(type="string", length=255, nullable=true)
-     * @Groups({"read","readUser","results","write"})
+     * @Groups({"read","readUser","results","write","readRelayPoint"})
      */
     private $venue;
 
@@ -295,7 +298,7 @@ class Address implements \JsonSerializable
     /**
      * @var array|null Label for display
      *
-     * @Groups({"read","readUser","readCommunity","readEvent","results","pt"})
+     * @Groups({"read","readUser","readCommunity","readEvent","results","pt","readRelayPoint"})
      */
     private $displayLabel;
 
@@ -304,6 +307,7 @@ class Address implements \JsonSerializable
      *
      * @ORM\OneToOne(targetEntity="App\RelayPoint\Entity\RelayPoint", mappedBy="address")
      * @Groups({"read","pt"})
+     * @MaxDepth(1)
      */
     private $relayPoint;
 
@@ -363,7 +367,7 @@ class Address implements \JsonSerializable
     /**
      * @var string|null Icon fileName.
      *
-     * @Groups({"read"})
+     * @Groups({"read","readRelayPoint"})
      */
     private $icon;